requirements.txt
asgiref==3.7.2
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.40.0
Django==4.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==23.0.0
idna==3.11
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.32.5
six==1.17.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.5.0
whitenoise==6.6.0
